requirements.txt
transformers==4.28.0
Pillow==9.3.0
torch==2.0.1
timm==1.0.3
